refactor(galgame): split rapidocr_support.py into 4 sub-modules
Move paths, model registry, runtime loading, and inspect/download out of
rapidocr_support.py into separate _* modules. Re-export everything from
rapidocr_support.py to preserve public API used by ocr_reader, service,
study_companion, and tests.

Uses deferred imports to break a circular dep between _paths and
_model_registry (RAPIDOCR_PACKAGE_NAME) and between _model_registry and
_runtime (_rapidocr_package_dir).

Pure code-move + re-export, no logic or formatting changes.
